tecolotl-nucleo/src/main/java/tecolotl/nucleo/persistencia/entidad/NivelLenguajeEntidad.java
package tecolotl.nucleo.persistencia.entidad;

import javax.persistence.*;

@Entity
@Table(name = "nivel_lenguaje", schema = "nucleo")
@NamedQuery(name = "NivelLenguaje.busca", query = "SELECT n FROM NivelLenguajeEntidad n")
public class NivelLenguajeEntidad {

    private short id;
    private String nombre;
    private String descripcion;

    @Id
    @Column(name = "id")
    public short getId() {
        return id;
    }

    public void setId(short id) {
        this.id = id;
    }

    @Basic
    @Column(name = "nombre")
    public String getNombre() {
        return nombre;
    }

    public void setNombre(String nombre) {
        this.nombre = nombre;
    }

    @Basic
    @Column(name = "descripcion")
    public String getDescripcion() {
        return descripcion;
    }

    public void setDescripcion(String descripcion) {
        this.descripcion = descripcion;
    }

    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (o == null || getClass() != o.getClass()) return false;

        NivelLenguajeEntidad that = (NivelLenguajeEntidad) o;

        if (id != that.id) return false;
        if (nombre != null ? !nombre.equals(that.nombre) : that.nombre != null) return false;
        if (descripcion != null ? !descripcion.equals(that.descripcion) : that.descripcion != null) return false;

        return true;
    }

    @Override
    public int hashCode() {
        int result = (int) id;
        result = 31 * result + (nombre != null ? nombre.hashCode() : 0);
        result = 31 * result + (descripcion != null ? descripcion.hashCode() : 0);
        return result;
    }
}
